pyproject.toml
[tool.poetry]
name = "drf-auth"
version = "0.1.0"
description = ""
authors = ["Vijayeta Rangarajan <vijayetar@gmail.com>"]

[tool.poetry.dependencies]
python = "~3.8"
django = "^3.0.7"
djangorestframework = "^3.11.0"
djangorestframework_simplejwt = "^4.4.0"
whitenoise = "^5.1.0"
django-cors-headers = "^3.4.0"
gunicorn = "^20.0.4"
psycopg2-binary = "^2.8.5"

[tool.poetry.dev-dependencies]
black = "^19.10b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
